Fixed updating appeals

+++ resources/js/components/Appeals.vue
@@ -292,20 +292,23 @@ export default {
                     this.currentAppeal,
                     { headers: { Authorization: `Bearer ${token}` } }
                 );
+                window.location.reload();
                 console.log("Ответ сервера:", response.data);
                 this.showModal = false;
             } catch (error) {
                 console.error("Ошибка при создании жалобы:", error.response?.data || error);
             }
         },
-
         async deleteAppeal(id) {
+            const isConfirmed = confirm('Вы уверены, что хотите удалить эту запись?');
+            if (!isConfirmed) return;
+
             try {
                 const token = localStorage.getItem('authToken');
                 await axios.delete(`http://localhost:8000/api/appeals/${id}`, {
                     headers: { Authorization: `Bearer ${token}` },
                 });
-                this.fetchAppeals(); // Обновление списка жалоб
+                this.fetchAppeals();
             } catch (error) {
                 console.error('Ошибка при удалении жалобы:', error);
             }
@@ -331,20 +334,6 @@ export default {
                 this.users = response.data.data; // Предполагаем, что API возвращает массив ролей
             } catch (error) {
                 console.error('Ошибка при загрузке ролей:', error);
-            }
-        },
-        async deleteAppeal(id) {
-            const isConfirmed = confirm('Вы уверены, что хотите удалить эту запись?');
-            if (!isConfirmed) return;
-
-            try {
-                const token = localStorage.getItem('authToken');
-                await axios.delete(`http://localhost:8000/api/appeals/${id}`, {
-                    headers: { Authorization: `Bearer ${token}` },
-                });
-                this.fetchAppeals();
-            } catch (error) {
-                console.error('Ошибка при удалении жалобы:', error);
             }
         },
         statusClass(status) {
